requirements.txt
flirimageextractor
PyExifTool
gdal==3.6.2